+++ apps/web/app/(main)/layout.tsx
@@ -1,10 +1,10 @@
 import type { ReactNode } from 'react';
 
-import { getAuthServerSide } from '@/lib/services/auth';
+import { auth } from '@/features/auth/actions/server';
 import { MainLayoutClient } from './_layout.client';
 
 const MainLayout = async ({ children }: { children: ReactNode }) => {
-  const { user } = await getAuthServerSide();
+  const { user } = await auth();
 
   return <MainLayoutClient user={user}>{children}</MainLayoutClient>;
 };
