+++ apps/web/app/(main)/my-space/logs/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { Post } from '@/components/post';
-import { getPosts } from '@/lib/actions/posts/get-posts';
+import { Post } from '@/features/post/components';
+import { getPosts } from '@/features/post/actions/get-posts';
 import { useQuery } from '@tanstack/react-query';
 
 const MyLogsPage = () => {
