+++ apps/web/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { LoginForm } from '@/components/auth/login-form';
+import { LoginForm } from '@/features/auth/components/login-form';
 
 const LoginPage = () => {
   return <LoginForm />;
