+++ apps/web/app/auth/register/page.tsx
@@ -1,4 +1,4 @@
-import { RegisterForm } from '@/components/auth/register-form';
+import { RegisterForm } from '@/features/auth/components/register-form';
 
 const RegisterPage = () => {
   return <RegisterForm />;
