+++ apps/web/features/auth/actions/server/index.ts
@@ -1,0 +1,4 @@
+import { auth } from './get-auth';
+import { verifyAuth } from './verify-auth';
+
+export { auth, verifyAuth };
