+++ apps/web/features/auth/components/register-form.tsx
@@ -19,35 +19,49 @@ import {
   FormLabel,
   FormMessage,
 } from '@workspace/ui/components/form';
-import { loginSchema, type LoginFormData } from '@workspace/schemas';
-import { useAuthStore } from '@/stores/auth-store';
+import { registerSchema, type RegisterFormData } from '@workspace/schemas';
+import { signup } from '@/features/auth/actions/client';
 
-export const LoginForm = ({ className, ...props }: React.ComponentProps<'form'>) => {
+export const RegisterForm = ({ className, ...props }: React.ComponentProps<'form'>) => {
   const router = useRouter();
   const searchParams = useSearchParams();
-  const login = useAuthStore((state) => state.login);
 
   const [isPending, startTransition] = useTransition();
 
   const callbackUrl = searchParams.get('callbackUrl') || '/feed';
 
-  const form = useForm<LoginFormData>({
-    resolver: zodResolver(loginSchema),
+  const form = useForm<RegisterFormData>({
+    resolver: zodResolver(registerSchema),
     defaultValues: {
+      name: '',
       email: '',
       password: '',
+      confirmPassword: '',
+      terms: false,
     },
   });
 
-  const onSubmit = async (data: LoginFormData) => {
+  const onSubmit = (data: RegisterFormData) => {
+    const { name, email, password, confirmPassword, terms } = data;
+    if (password !== confirmPassword) {
+      toast.error("Passwords don't match");
+      return;
+    }
+
+    if (!terms) {
+      toast.error('Agree to the terms and conditions to signup');
+      return;
+    }
+
     startTransition(async () => {
       try {
-        await login(data);
+        await signup({ name, email, password });
+        toast.success('Signup successful');
 
         // Redirect to callback url
         router.push(decodeURIComponent(callbackUrl));
       } catch (error) {
-        console.error('Login failed:', error);
+        console.error('Signup failed:', error);
         toast.error('Something went wrong!');
       }
     });
@@ -61,13 +75,27 @@ export const LoginForm = ({ className, ...props }: React.ComponentProps<'form'>)
         {...props}
       >
         <div className="flex flex-col items-center gap-2 text-center">
-          <h1 className="text-2xl font-bold">Login to your account</h1>
+          <h1 className="text-2xl font-bold">Create an account</h1>
           <p className="text-muted-foreground text-sm text-balance">
-            Enter your email below to login to your account
+            Enter your information below to create your account
           </p>
         </div>
 
         <div className="grid gap-6">
+          <FormField
+            control={form.control}
+            name="name"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Full Name</FormLabel>
+                <FormControl>
+                  <Input type="text" placeholder="John Doe" disabled={isPending} {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
           <FormField
             control={form.control}
             name="email"
@@ -87,16 +115,7 @@ export const LoginForm = ({ className, ...props }: React.ComponentProps<'form'>)
             name="password"
             render={({ field }) => (
               <FormItem>
-                <div className="flex items-center">
-                  <FormLabel>Password</FormLabel>
-                  <Link
-                    href="/"
-                    className="ml-auto text-sm underline-offset-4 hover:underline"
-                    tabIndex={-1}
-                  >
-                    Forgot your password?
-                  </Link>
-                </div>
+                <FormLabel>Password</FormLabel>
                 <FormControl>
                   <Input type="password" disabled={isPending} {...field} />
                 </FormControl>
@@ -105,9 +124,51 @@ export const LoginForm = ({ className, ...props }: React.ComponentProps<'form'>)
             )}
           />
 
+          <FormField
+            control={form.control}
+            name="confirmPassword"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Confirm Password</FormLabel>
+                <FormControl>
+                  <Input type="password" disabled={isPending} {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
+          <FormField
+            control={form.control}
+            name="terms"
+            render={({ field }) => (
+              <FormItem className="flex items-center space-x-2">
+                <FormControl>
+                  <input
+                    type="checkbox"
+                    className="h-4 w-4 rounded border-gray-300"
+                    checked={field.value}
+                    onChange={field.onChange}
+                  />
+                </FormControl>
+                <FormLabel className="text-sm">
+                  I agree to the{' '}
+                  <Link href="/" className="underline underline-offset-4">
+                    Terms of Service
+                  </Link>{' '}
+                  and{' '}
+                  <Link href="/" className="underline underline-offset-4">
+                    Privacy Policy
+                  </Link>
+                </FormLabel>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+
           <Button type="submit" className="w-full" disabled={isPending}>
             {isPending && <Loader2 className="size-4 animate-spin" />}
-            Login
+            Create Account
           </Button>
 
           <div className="after:border-border relative text-center text-sm after:absolute after:inset-0 after:top-1/2 after:z-0 after:flex after:items-center after:border-t">
@@ -123,14 +184,14 @@ export const LoginForm = ({ className, ...props }: React.ComponentProps<'form'>)
                 fill="currentColor"
               />
             </svg>
-            Login with GitHub
+            Sign up with GitHub
           </Button>
         </div>
 
         <div className="text-center text-sm">
-          Don&apos;t have an account?{' '}
-          <Link href="/auth/register" className="underline underline-offset-4">
-            Sign up
+          Already have an account?{' '}
+          <Link href="/auth/login" className="underline underline-offset-4">
+            Sign in
           </Link>
         </div>
       </form>
